+++ packages/authentication/modules/user/useJWTUser/types.ts
@@ -1,3 +1,5 @@
+import { ServerSideRenderingOption } from '@baseapp-frontend/utils'
+
 import { UseQueryOptions } from '@tanstack/react-query'
 
 import UserApi from '../../../services/user'
@@ -5,7 +7,7 @@ import { ICookieName } from '../../../types/auth'
 
 type ApiClass = Pick<typeof UserApi, 'getUser'>
 
-export interface IUseJWTUser<IUser> extends ICookieName {
+export interface IUseJWTUser<IUser> extends ICookieName, ServerSideRenderingOption {
   options?: UseQueryOptions<IUser, unknown, IUser, any>
   ApiClass?: ApiClass
 }