+++ packages/utils/functions/cookie/types.ts
@@ -1,0 +1,5 @@
+import { ServerSideRenderingOption } from '../../types/server'
+
+export interface GetCookieOptions extends ServerSideRenderingOption {
+  parseJSON?: boolean
+}