+++ packages/utils/types/server.ts
@@ -1,0 +1,3 @@
+export type ServerSideRenderingOption = {
+  noSSR?: boolean
+}